refactor: rename modules extract to file and batch to sql
Module names now reflect the data source rather than direction.
