+++ heave/__init__.py
@@ -1,2 +1,2 @@
 """Root of the heave package."""
-from .extract import Table
+from .file import Table
