+++ tests/test_sql.py
@@ -1,7 +1,7 @@
-"""Tests for the batch module."""
+"""Tests for the sql module."""
 from sqlalchemy import text
 
-from heave import Table, batch
+from heave import Table, sql
 
 
 class TestSql:
@@ -9,7 +9,7 @@ class TestSql:
 
     def test_reflect_table(self, connection):
         """Test the reflect_table function."""
-        table = batch.reflect_table(connection, "user")
+        table = sql.reflect_table(connection, "user")
         assert table.name == "user"
 
     def test_insert(self, connection):
@@ -20,8 +20,8 @@ class TestSql:
                 ("jane.doe", "janedoe@example.com", "yourSecurePassword"),
             ]
         )
-        sql_table = batch.reflect_table(connection, "user")
-        batch.insert(connection, sql_table, data)
+        sql_table = sql.reflect_table(connection, "user")
+        sql.insert(connection, sql_table, data)
         result = connection.execute(
             text("SELECT * FROM user WHERE username = 'jane.doe';")
         )
